[compose-server] feat : DB용 레포지토리 정리
