+++ core/data/src/main/java/com/hegunhee/newsimplememoapp/data/di/DefaultRepositoryModule.kt
@@ -2,10 +2,8 @@ package com.hegunhee.newsimplememoapp.data.di
 
 import com.hegunhee.newsimplememoapp.data.repository.DefaultCategoryRepository
 import com.hegunhee.newsimplememoapp.data.repository.DefaultMemoRepository
-import com.hegunhee.newsimplememoapp.data.repository.DefaultMemoTempRepository
 import com.hegunhee.newsimplememoapp.domain.repository.CategoryRepository
 import com.hegunhee.newsimplememoapp.domain.repository.MemoRepository
-import com.hegunhee.newsimplememoapp.domain.repository.MemoTempRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -19,19 +17,13 @@ abstract class DefaultRepositoryModule {
 
     @Singleton
     @Binds
-    abstract fun provideDefaultRepository(
-        defaultMemoRepository: DefaultMemoRepository
-    ) : MemoRepository
-
-    @Singleton
-    @Binds
     abstract fun provideDefaultCategoryRepository(
         defaultCategoryRepository: DefaultCategoryRepository
     ) : CategoryRepository
 
     @Singleton
     @Binds
-    abstract fun provideDefaultMemoTempRepository(
-        defaultMemoTempRepository: DefaultMemoTempRepository
-    ) : MemoTempRepository
+    abstract fun provideDefaultMemoRepository(
+        defaultMemoTempRepository: DefaultMemoRepository
+    ) : MemoRepository
 }
