+++ core/domain/src/main/java/com/hegunhee/newsimplememoapp/domain/usecase/memo/DeleteMemoUseCase.kt
@@ -1,13 +1,13 @@
 package com.hegunhee.newsimplememoapp.domain.usecase.memo
 
-import com.hegunhee.newsimplememoapp.domain.repository.MemoTempRepository
+import com.hegunhee.newsimplememoapp.domain.repository.MemoRepository
 import javax.inject.Inject
 
 class DeleteMemoUseCase @Inject constructor(
-    private val memoTempRepository: MemoTempRepository
+    private val memoRepository: MemoRepository
 ) {
 
     suspend operator fun invoke(memoId: Int) : Result<Int> {
-        return memoTempRepository.deleteMemo(memoId)
+        return memoRepository.deleteMemo(memoId)
     }
 }
