+++ core/domain/src/main/java/com/hegunhee/newsimplememoapp/domain/usecase/memo/GetMemoUseCase.kt
@@ -1,14 +1,14 @@
 package com.hegunhee.newsimplememoapp.domain.usecase.memo
 
 import com.hegunhee.newsimplememoapp.domain.model.memo.Memo
-import com.hegunhee.newsimplememoapp.domain.repository.MemoTempRepository
+import com.hegunhee.newsimplememoapp.domain.repository.MemoRepository
 import javax.inject.Inject
 
 class GetMemoUseCase @Inject constructor(
-    private val memoTempRepository: MemoTempRepository
+    private val memoRepository: MemoRepository
 ) {
 
     suspend operator fun invoke(memoId : Int) : Result<Memo> {
-        return memoTempRepository.getMemo(memoId)
+        return memoRepository.getMemo(memoId)
     }
 }
