+++ core/domain/src/main/java/com/hegunhee/newsimplememoapp/domain/usecase/memo/GetMemosSummaryUseCase.kt
@@ -1,11 +1,11 @@
 package com.hegunhee.newsimplememoapp.domain.usecase.memo
 
 import com.hegunhee.newsimplememoapp.domain.model.memo.MemosSummary
-import com.hegunhee.newsimplememoapp.domain.repository.MemoTempRepository
+import com.hegunhee.newsimplememoapp.domain.repository.MemoRepository
 import javax.inject.Inject
 
 class GetMemosSummaryUseCase @Inject constructor(
-    private val tempRepository: MemoTempRepository
+    private val tempRepository: MemoRepository
 ) {
 
     suspend operator fun invoke(year : Int,month : Int) : Result<MemosSummary> {
