+++ core/domain/src/main/java/com/hegunhee/newsimplememoapp/domain/usecase/memo/GetStaticsMemosUseCase.kt
@@ -2,14 +2,14 @@ package com.hegunhee.newsimplememoapp.domain.usecase.memo
 
 import com.hegunhee.newsimplememoapp.domain.model.memo.IncomeExpenseType
 import com.hegunhee.newsimplememoapp.domain.model.memo.StaticsMemos
-import com.hegunhee.newsimplememoapp.domain.repository.MemoTempRepository
+import com.hegunhee.newsimplememoapp.domain.repository.MemoRepository
 import javax.inject.Inject
 
 class GetStaticsMemosUseCase @Inject constructor(
-    private val memoTempRepository: MemoTempRepository
+    private val memoRepository: MemoRepository
 ) {
 
     suspend operator fun invoke(incomeExpenseType: IncomeExpenseType,year : Int,month : Int) : Result<StaticsMemos> {
-        return memoTempRepository.getStaticsMemo(incomeExpenseType,year,month)
+        return memoRepository.getStaticsMemo(incomeExpenseType,year,month)
     }
 }
