+++ core/domain/src/main/java/com/hegunhee/newsimplememoapp/domain/usecase/memo/UpdateMemoUseCase.kt
@@ -1,14 +1,14 @@
 package com.hegunhee.newsimplememoapp.domain.usecase.memo
 
 import com.hegunhee.newsimplememoapp.domain.model.memo.MemoForm
-import com.hegunhee.newsimplememoapp.domain.repository.MemoTempRepository
+import com.hegunhee.newsimplememoapp.domain.repository.MemoRepository
 import javax.inject.Inject
 
 class UpdateMemoUseCase @Inject constructor(
-    private val memoTempRepository: MemoTempRepository
+    private val memoRepository: MemoRepository
 ) {
 
     suspend operator fun invoke(memoId : Int,memoForm : MemoForm) : Result<Int>{
-        return memoTempRepository.updateMemo(memoId,memoForm)
+        return memoRepository.updateMemo(memoId,memoForm)
     }
 }
